refactor : useComments 폴더 변경
